Fleshed out REST API endpoints

+++ src/SimU-GameService.Contracts/Requests/CreateAgentRequest.cs
@@ -4,4 +4,4 @@ public record CreateAgentRequest(
     string Username,
     string Description,
     Guid CreatorId,
-    decimal IncubationTimeInHours);
+    float IncubationTimeInHours);
